feat: updating classes for seller rating feature

+++ backend/src/main/java/divar/service/impl/FavoriteServiceImpl.java
@@ -77,7 +77,7 @@ public class FavoriteServiceImpl implements FavoriteService {
                     response.setId(advertisement.getId());
                     response.setTitle(advertisement.getTitle());
                     response.setDescription(advertisement.getDescription());
-                    response.setPrice((long) advertisement.getPrice());
+                    response.setPrice(advertisement.getPrice());
 
                     response.setStatus(advertisement.getStatus());
